refactored code to make review score dynamic
