+++ src/pages/shoppage/sections/shopdetails/ShopDetails.jsx
@@ -6,24 +6,24 @@ import {useEffect, useState} from "react";
 import axios from "axios";
 import {BASE_URL} from "@/utils/urlBuilder.js";
 
-function ShopDetails() {
-    const { id } = useParams();
-    const [shop, setShop] = useState({});
-
-    useEffect(() => {
-        async function fetchShop() {
-            try{
-                const response = await axios.get(BASE_URL + `shops/${id}/profile`);
-                setShop(response.data);
-                console.log(response.data);
-            } catch(e) {
-                console.error(e);
-            }
-        }
-
-        fetchShop();
-
-    }, [])
+function ShopDetails({ shop }) {
+    // const { id } = useParams();
+    // const [shop, setShop] = useState({});
+    //
+    // useEffect(() => {
+    //     async function fetchShop() {
+    //         try{
+    //             const response = await axios.get(BASE_URL + `shops/${id}/profile`);
+    //             setShop(response.data);
+    //             console.log(response.data);
+    //         } catch(e) {
+    //             console.error(e);
+    //         }
+    //     }
+    //
+    //     fetchShop();
+    //
+    // }, [])
 
 
     return (
@@ -34,7 +34,7 @@ function ShopDetails() {
             <div className="shop-details__content">
                 <ShopInfo
                     shopName={shop.name}
-                    rating={4}
+                    rating={shop.averageRating}
                     reviewCount={shop.numberOfReviews}
                 ></ShopInfo>
                 <p className="shop-details__description">{shop.description}</p>
